passwords/requirements.txt
pylint
pytest
psycopg2-binary
python-dotenv
argon2-cffi